Add files via upload
The q(t) & q_dot(t) time histories are used for as training data for SINDy to identify the non-linear wake oscillator model. The discovered model is then used as a coupled ODE system together along with the structural ODE. The responses from the SINDy identified system are analyzed and compared with the responses from the original system.